requirements.txt
flask~=2.1.1
requests~=2.27.1
python-binance
pandas~=1.4.2
SQLAlchemy~=1.4.37